workmail_authorizer/requirements.txt
pyjwt>=2.0,<3.0
boto3==1.35.80
botocore==1.35.80